cerulean_cloud/cloud_run_sea_ice/requirements.txt
fastapi
uvicorn[standard]
requests
google-cloud-storage
